Invoice Data Extraction and Verification from Scanned PDFs/requirements.txt
# Core PDF and Image Processing
pdf2image==1.17.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0

# OCR Engines
pytesseract==0.3.10
easyocr==1.7.0
torch>=1.13.1
torchvision>=0.14.1

# Data Processing and Output
pandas==2.0.3
openpyxl==3.1.2

# Text Processing and Pattern Matching
regex==2023.8.8
rapidfuzz==3.5.2  # replaced fuzzywuzzy + Levenshtein

# Image Enhancement and Preprocessing
scipy==1.11.2
imutils==0.5.4

# Configuration and Utilities
pyyaml==6.0.1
click==8.1.7

# Testing (move to dev-requirements.txt if needed)
pytest==7.4.2
pytest-cov==4.1.0

# Logging and Progress
tqdm==4.66.1
colorlog==6.7.0

# Optional: For advanced image processing
deskew==1.3.2
textdistance==4.6.0
